Create README.md
This Repo will have all the practical things/repos/sub-repos you'll need to practically create, in order to understand Git. 
This isn't something scary. When you don't know a thing or the terminology, is only then when you get scared of what actually is happening.

No worries!
This documentation is written in a normal human friendly vibes, you won't get lost in the terminoogies or any other wy.
And definitely, this is going to make you fall in love with learning.
So, No worries, and read the documentation, it'll expand your understanding and knowledge in a friendly humanising way.